+++ test/cases/transaction_test_sqlserver.rb
@@ -30,6 +30,23 @@ class TransactionTestSQLServer < ActiveRecord::TestCase
     end
   end
 
+  it 'can use an isolation level and reverts back to starting isolation level' do
+    begin
+      in_level = nil
+      begin_level = connection.user_options_isolation_level
+      begin_level.must_match %r{read committed}
+      Ship.transaction(isolation: :serializable) do
+        Ship.create! name: 'Black Pearl'
+        in_level = connection.user_options_isolation_level
+      end
+      after_level = connection.user_options_isolation_level
+      in_level.must_match %r{serializable}i
+      after_level.must_match %r{read committed}
+    ensure
+      connection.set_transaction_isolation_level 'READ COMMITTED'
+    end
+  end
+
 
   protected